+++ index.js
@@ -7,7 +7,8 @@ module.exports = {
     'no-loops': require('./rules/no-loops'),
     'no-mutating-assign': require('./rules/no-mutating-assign'),
     'no-mutation': require('./rules/no-mutation'),
-    'no-this': require('./rules/no-this')
+    'no-this': require('./rules/no-this'),
+    'no-throw': require('./rules/no-throw')
   },
   configs: {
     recommended: {
@@ -18,7 +19,8 @@ module.exports = {
         'fp/no-loops': 'error',
         'fp/no-mutating-assign': 'error',
         'fp/no-mutation': 'error',
-        'fp/no-this': 'error'
+        'fp/no-this': 'error',
+        'fp/no-throw': 'error'
       }
     }
   }