+++ rules/no-throw.js
@@ -1,0 +1,12 @@
+'use strict';
+
+module.exports = function (context) {
+  return {
+    ThrowStatement: function (node) {
+      context.report({
+        node: node,
+        message: 'Unallowed use of `throw`'
+      });
+    }
+  };
+};